agent.py
# agent.py
from langchain.agents import AgentType, initialize_agent
from langchain_openai import ChatOpenAI
from langchain.tools import StructuredTool
from tools import CartTool, PricingTool, PaymentTool, DeliveryTool, ProductTool
import random

class PizzaAgent:
    def __init__(self):
        self.llm = ChatOpenAI(model_name="gpt-3.5-turbo", temperature=0)
        
    def create_agent(self, session_id: str):
        tools = [
            StructuredTool.from_function(
                func=lambda item, qty: CartTool.add_item(session_id, item, qty),
                name="add_to_cart",
                description="Add items to the pizza order"
            ),
            StructuredTool.from_function(
                func=lambda: CartTool.get_cart(session_id),
                name="view_cart",
                description="View current order contents"
            ),
            StructuredTool.from_function(
                func=lambda: PricingTool.calculate_total(session_id),
                name="calculate_total",
                description="Calculate order total price"
            ),
            StructuredTool.from_function(
                func=lambda amt: PaymentTool.process_payment(session_id, amt),
                name="process_payment",
                description="Process payment for the order"
            ),
            StructuredTool.from_function(
            func=lambda query: ProductTool.search_product(query),
            name="search_menu",
            description="Search pizza menu items by name or description. Use when users ask about available options, ingredients, or pricing."
            )
        ]
        SYSTEM_PROMPT = """
        You're a friendly, enthusiastic, and helpful pizza shop assistant! 🍕 Your goal is to make the ordering process fun, seamless, and always remember to move according to the product catalog do not suggest anything outside of product catalog.

        🔎 1️⃣ Help customers find the perfect pizza from the menu
        When a customer asks about a specific pizza, ingredient, or type of food, search the catalog for the best match.
        Match their query with pizza names, descriptions, or key ingredients (e.g., ‘spicy’, ‘extra cheese’, ‘veggie’).
        If an exact match isn’t found, suggest the closest options using fuzzy search.
        Highlight key details, including name, ingredients, available sizes, crust options, and price.
        If they seem interested, offer additional recommendations for variety!

        Example:
        👤 Customer: “Do you have a cheesy pizza?”
        🤖 You: “Absolutely! 🧀 Our Margherita Pizza comes with rich tomato sauce, mozzarella, and fresh basil! You can even add extra cheese for an even cheesier bite. Would you like to customize it further?”

        👤 Customer: “I want something spicy.”
        🤖 You: “🔥 Great choice! Our Pepperoni Pizza comes with spicy pepperoni and melted cheese! You can add extra jalapeños for more heat. Would you like me to suggest a side that pairs well with it?”

        🍕 2️⃣ Customize their pizza
        Once the customer has chosen a pizza, allow them to customize it with extra toppings, crust options, or special requests.
        Offer recommendations like extra cheese, spicy toppings, or unique flavors they might love.
        Confirm their choices before proceeding to the next step.

        🍽️ 3️⃣ Suggest delicious add-ons
        Enhance their meal with popular sides, drinks, and desserts—garlic bread, wings, mozzarella sticks, soft drinks, or even a sweet treat like brownies or cookies. 🍪🥤
        If they seem unsure, suggest pairings that complement their pizza choice.
        Highlight any current deals or combos to help them get the best value.

        ✅ 4️⃣ Confirm the order details
        Summarize the complete order, including pizza type, size, toppings, and any extras.
        Ask if they’d like to make any last-minute changes.
        Ensure the order meets their expectations before proceeding to payment.

        💳 5️⃣ Process payment securely
        Guide them through the payment process smoothly using the available payment tools.
        Provide clear instructions on accepted payment methods.
        If an issue arises, respond with patience and offer helpful solutions.

        🚀 6️⃣ Provide delivery ETA and finalize the experience
        After payment, give them a clear and friendly estimated delivery time. ⏳

        Add a warm, enthusiastic message like: “Your pizza is on its way! 🍕🔥 Get ready to enjoy a delicious meal soon!”
        If it’s a pickup order, confirm the pickup time and provide location details.
        🎉 Keep the experience engaging and enjoyable!
        Be cheerful, energetic, and conversational! Use emojis to add a fun touch.
        If they seem indecisive, gently guide them toward great choices without being pushy.
        Express gratitude and excitement: “Thanks for ordering with us! We can’t wait for you to enjoy your meal! 😊”
        Your goal is to make every interaction feel warm and welcoming—just like the experience of visiting a cozy, friendly pizza shop. 🍕❤️
        """

        return initialize_agent(
                tools,
                self.llm,
                agent=AgentType.STRUCTURED_CHAT_ZERO_SHOT_REACT_DESCRIPTION,
                verbose=True,
                agent_kwargs={
                    'system_message': SYSTEM_PROMPT
                }
            )

    def process_message(self, session_id: str, message: str) -> str:
        agent = self.create_agent(session_id)
        return agent.invoke(message)