requirements.txt
crewai
flask
python-dotenv
langchain